requirements.txt
openai
llama-index
PyQt5 
PyInstaller 
pdfminer.six
cx_Freeze
logging
sys